# packages/api/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Directorio Telefónico",
    "description": "Una API RESTful para gestionar contactos en un directorio telefónico",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor local"
    }
  ],
  "paths": {
    "/api/contactos": {
      "get": {
        "summary": "Obtener todos los contactos",
        "responses": {
          "200": {
            "description": "Lista de contactos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactosResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear un nuevo contacto",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactoInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contacto creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactoResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/contactos/{id}": {
      "get": {
        "summary": "Obtener un contacto por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contacto encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactoResponse"
                }
              }
            }
          },
          "404": {
            "description": "Contacto no encontrado"
          }
        }
      },
      "put": {
        "summary": "Actualizar un contacto existente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactoInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contacto actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactoResponse"
                }
              }
            }
          },
          "404": {
            "description": "Contacto no encontrado"
          }
        }
      },
      "delete": {
        "summary": "Eliminar un contacto",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contacto eliminado exitosamente"
          },
          "404": {
            "description": "Contacto no encontrado"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Contacto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "nombre": {
            "type": "string"
          },
          "apellido": {
            "type": "string"
          },
          "telefono": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "direccion": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ContactoInput": {
        "type": "object",
        "required": ["nombre", "apellido", "telefono"],
        "properties": {
          "nombre": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "apellido": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "telefono": {
            "type": "string",
            "minLength": 9,
            "maxLength": 20
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "direccion": {
            "type": "string"
          }
        }
      },
      "ContactoResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/Contacto"
          }
        }
      },
      "ContactosResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Contacto"
            }
          }
        }
      }
    }
  }
} 